refactor: abstract out empty workspace components
