+++ frontend/src/features/workspace/components/EmptyWorkspace/EmptyDefaultWorkspace.tsx
@@ -1,0 +1,13 @@
+import { EmptyWorkspace, EmptyWorkspacePage } from './EmptyWorkspace'
+
+export const EmptyDefaultWorkspace = ({
+  isLoading,
+  handleOpenCreateFormModal,
+}: EmptyWorkspacePage) => (
+  <EmptyWorkspace
+    isLoading={isLoading}
+    handleOpenCreateFormModal={handleOpenCreateFormModal}
+    title={"You don't have any forms yet"}
+    subText={'Get started by creating a new form'}
+  />
+)
