+++ frontend/src/features/workspace/components/EmptyWorkspace/index.ts
@@ -1,1 +1,2 @@
-export { EmptyDefaultWorkspace, EmptyNewWorkspace } from './EmptyWorkspace'
+export { EmptyDefaultWorkspace } from './EmptyDefaultWorkspace'
+export { EmptyNewWorkspace } from './EmptyNewWorkspace'
